refactor: 테스트를 용이하기 위해 RacingCarController를 포트와 어댑터 패턴으로 추상화
