+++ src/main/java/racingcar/port/in/RacingCarInputPort.java
@@ -1,0 +1,8 @@
+package racingcar.port.in;
+
+import java.util.List;
+
+public interface RacingCarInputPort {
+    List<String> getCarNames();
+    int getTryCount();
+}
